Rename: 파일명 수정 CardType -> cardType
